# docs/openapi.yaml
openapi: 3.0.3
info:
  title: GoStudy Enterprise API
  version: 1.0.0
  description: |
    GoStudy 企业应用骨架的 HTTP 接口。除健康检查、注册和登录外，接口使用 Bearer JWT 鉴权。
    所有响应均包含业务码 `code`、提示信息 `message` 和可选数据字段 `data`。
servers:
  - url: http://127.0.0.1:8080
    description: 本地默认地址
tags:
  - name: 系统
  - name: 用户
  - name: 认证
  - name: 任务
  - name: 审计
paths:
  /health:
    get:
      tags: [系统]
      summary: 查询服务及依赖健康状态
      responses:
        '200':
          description: 服务可用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
  /api/v1/users:
    post:
      tags: [用户]
      summary: 注册用户
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterUserInput'
      responses:
        '201':
          description: 注册成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/InvalidInput'
        '409':
          $ref: '#/components/responses/Conflict'
    get:
      tags: [用户]
      summary: 查询用户列表
      description: 仅管理员角色可访问。
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPageResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/auth/login:
    post:
      tags: [认证]
      summary: 登录并签发 JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginInput'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/InvalidInput'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/auth/refresh:
    post:
      tags: [认证]
      summary: 轮换当前有效 JWT
      description: 当前 Token 会立即吊销，并返回新的 Token。
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 轮换成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/auth/logout:
    post:
      tags: [认证]
      summary: 吊销当前 JWT
      description: 吊销列表默认保存在内存，服务重启后会清空。
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 退出成功
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/me:
    get:
      tags: [用户]
      summary: 查询当前登录用户资料
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/audits:
    get:
      tags: [审计]
      summary: 查询审计记录列表
      description: 仅管理员角色可访问。
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditPageResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/tasks:
    post:
      tags: [任务]
      summary: 创建当前用户的任务
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskInput'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          $ref: '#/components/responses/InvalidInput'
        '401':
          $ref: '#/components/responses/Unauthorized'
    get:
      tags: [任务]
      summary: 查询当前用户的任务列表
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskPageResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/tasks/{id}/complete:
    patch:
      tags: [任务]
      summary: 完成当前用户的任务
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 完成成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    InvalidInput:
      description: 参数不合法
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example: { code: 40001, message: invalid input }
    Unauthorized:
      description: 未认证或 JWT 无效
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example: { code: 40101, message: missing bearer token }
    Forbidden:
      description: 无权访问资源
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example: { code: 40301, message: insufficient permission }
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example: { code: 40401, message: resource not found }
    Conflict:
      description: 资源冲突
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example: { code: 40901, message: resource conflict }
  schemas:
    RegisterUserInput:
      type: object
      required: [name, email, password]
      properties:
        name: { type: string, example: Alice }
        email: { type: string, format: email, example: alice@example.com }
        password: { type: string, format: password, minLength: 6, example: secret1 }
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email, example: alice@example.com }
        password: { type: string, format: password, example: secret1 }
    CreateTaskInput:
      type: object
      required: [title]
      properties:
        title: { type: string, example: 发布企业骨架 }
    User:
      type: object
      properties:
        id: { type: integer, format: int64, example: 1 }
        name: { type: string, example: Alice }
        email: { type: string, format: email, example: alice@example.com }
        role: { type: string, enum: [user, admin], example: user }
        created_at: { type: string, format: date-time }
    Task:
      type: object
      properties:
        id: { type: integer, format: int64, example: 1 }
        title: { type: string, example: 发布企业骨架 }
        owner_id: { type: integer, format: int64, example: 1 }
        status: { type: string, enum: [pending, done], example: pending }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    AuditEntry:
      type: object
      properties:
        action: { type: string, example: task.created }
        actor_id: { type: integer, format: int64, example: 1 }
        resource: { type: string, example: task:1 }
        created_at: { type: string, format: date-time }
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code: { type: integer, example: 40001 }
        message: { type: string, example: invalid input }
    UserResponse:
      type: object
      properties:
        code: { type: integer, example: 0 }
        message: { type: string, example: ok }
        data: { $ref: '#/components/schemas/User' }
    TaskResponse:
      type: object
      properties:
        code: { type: integer, example: 0 }
        message: { type: string, example: ok }
        data: { $ref: '#/components/schemas/Task' }
    LoginResponse:
      type: object
      properties:
        code: { type: integer, example: 0 }
        message: { type: string, example: ok }
        data:
          type: object
          properties:
            token: { type: string, description: HS256 JWT }
            user: { $ref: '#/components/schemas/User' }
    TokenResponse:
      type: object
      properties:
        code: { type: integer, example: 0 }
        message: { type: string, example: ok }
        data:
          type: object
          properties:
            token: { type: string, description: 轮换后的 HS256 JWT }
    UserPageResponse:
      type: object
      properties:
        code: { type: integer, example: 0 }
        message: { type: string, example: ok }
        data:
          type: object
          properties:
            page: { type: integer, example: 1 }
            page_size: { type: integer, example: 20 }
            total: { type: integer, example: 1 }
            items:
              type: array
              items: { $ref: '#/components/schemas/User' }
    TaskPageResponse:
      type: object
      properties:
        code: { type: integer, example: 0 }
        message: { type: string, example: ok }
        data:
          type: object
          properties:
            page: { type: integer, example: 1 }
            page_size: { type: integer, example: 20 }
            total: { type: integer, example: 1 }
            items:
              type: array
              items: { $ref: '#/components/schemas/Task' }
    AuditPageResponse:
      type: object
      properties:
        code: { type: integer, example: 0 }
        message: { type: string, example: ok }
        data:
          type: object
          properties:
            page: { type: integer, example: 1 }
            page_size: { type: integer, example: 20 }
            total: { type: integer, example: 1 }
            items:
              type: array
              items: { $ref: '#/components/schemas/AuditEntry' }
    HealthResponse:
      type: object
      properties:
        code: { type: integer, example: 0 }
        message: { type: string, example: ok }
        data:
          type: object
          properties:
            name: { type: string, example: GoStudy Enterprise }
            env: { type: string, example: dev }
            dependencies:
              type: object
              additionalProperties:
                type: string
              example: { storage: skipped, cache: skipped, mq: skipped }